[!][update] remove decode/inst_16 and decode/inst_32 directory
