+++ src/decode.rs
@@ -3,6 +3,16 @@
 mod inst_16;
 mod inst_32;
 
+mod a_extension;
+mod base_i;
+mod c_extension;
+mod m_extension;
+mod priv_extension;
+mod zicfiss_extension;
+mod zicntr_extension;
+mod zicsr_extension;
+mod zifencei_extension;
+
 use crate::instruction::{Instruction, Opcode, OpcodeKind};
 use crate::{Extensions, Isa};
